pyproject.toml
[tool.poetry]
name = "bazel-pycross-poetry-gazelle-example"
version = "0.0.0"
description = ""
authors = [""]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "3.0.2"
python-dateutil = "2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
